refactor: 공연 목록 조회 시 장소 정보 조회 맵핑 코드 리팩토링
